scripts/requirements.txt
biopython>=1.67
